refactor: move fielding-view to components/softball/session-views/
